test rest controleur article

Modification des entitles pour inclure les relations entre commande bon article.

Point a verifier si tous les id des entité sont genere alétoirement

+++ src/main/java/com/gestionets/Gestion/Entreprise/Service/ArticleService.java
@@ -13,7 +13,7 @@ public interface ArticleService {
 
     void ModifierArticle(Article article);
 
-    void SupprimerArticle(Integer ID);
+    void SupprimerArticle(Long ID);
 
-    Optional<Article> RechercheArticleparID(Integer ID);
+    Optional<Article> RechercheArticleparID(Long ID);
 }
